asasas

+++ proj-java/src/main/java/engine/UserBD.java
@@ -15,41 +15,77 @@ public class UserBD {
 
     private HashMap<Long,User> usermap;
 
+    /**
+     * Construtor vazio
+     */
     public UserBD(){
         this.usermap = new HashMap<Long,User>();
     }
 
-    public UserBD(HashMap<Integer,User> u){
-        HashMap<Integer,User> usermap = new HashMap<>();
+    /**
+     * Construtor parametrizado
+     * @param HashMap<Long,User> 
+     */
+    public UserBD(HashMap<Long,User> u){
+        HashMap<Long,User> usermap = new HashMap<>();
         u.entrySet().stream().forEach(entry -> usermap.put(entry.getKey(), entry.getValue().clone()));
     }
 
+    /**
+     * Construtor de cópia
+     * @param UserBD
+     */
     public UserBD(UserBD u){
         usermap = u.getUserMap();
     }
 
+    /**
+     * Função que limpa a estrutura
+     */    
     public void clear(){
         this.usermap.clear();
     }
 
+    /**
+     * Dado um id, devolve o user correspondente
+     * @param long id
+     * @return User
+     */
     public User getUser(long id){
         return this.usermap.get(id);
     }
     
+    /**
+     * Dado um user, adiciona-o à base de dados
+     * @param User
+     */
     public void addUser(User u){
         this.usermap.put(u.getUserId(),u.clone());
     }
     
+    /**
+     * Devolve o map que contém todos os users e respetivos id's.
+     * @return HashMap<Long,User>
+     */
     public HashMap<Long,User> getUserMap(){
         HashMap<Long,User> usermap = new HashMap<>();
         this.usermap.entrySet().stream().forEach(entry -> usermap.put(entry.getKey(), entry.getValue().clone()));
         return usermap;
     }
     
+    /**
+     * Devolve o clone do objeto
+     * @return UserBD
+     */
     public UserBD clone(){
         return new UserBD(this);
     }
     
+    /**
+     * Verifica se dois objetos são iguais
+     * @param Object
+     * @return boolean
+     */
     public boolean equals(Object o){
         if(this==o) return true;
